+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Development
 
 * Handle singular collection names like `have(n).item` ([#18](https://github.com/yujinakayama/transpec/issues/18))
+* Handle collection accessors with arguments like `have(n).errors_on(...)` ([#18](https://github.com/yujinakayama/transpec/issues/18))
 
 ## v1.2.2